% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsurvdef.R
\name{plotsurvdef}
\alias{plotsurvdef}
\title{Plots two survival distributions.}
\usage{
plotsurvdef(survdefC, survdefT, xupper)
}
\arguments{
\item{survdefC}{the survival distribution of the control group (will be
plotted as a solid line), as a list in the form output by \code{survdef}.}

\item{survdefT}{the survival distribution of the control group (will be
plotted as a dashed line), as a list in the form output by \code{survdef}.}

\item{xupper}{the upper x axis limit for the plot.}
}
\description{
Plots two survival distributions.
}
\examples{
RIC<-survdef(times = 18, surv = .6)
MAC<-survdef(times = 3, haz=c(4.375*RIC$h(1),RIC$h(1)))
plotsurvdef(RIC, MAC, 24)
}
