% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSTdesign-package.R
\docType{package}
\name{RMSTdesign-package}
\alias{RMSTdesign}
\alias{RMSTdesign-package}
\title{RMSTdesign: Sample Size and Power Calculation for the Difference in Restricted Mean Survival Time}
\description{
Calculates the asymptotic power of the test of the difference in 
    restricted mean survival time under user-specified trial design parameters, 
    or the sample size required for a specified power. It also provides 
    the probability restricted mean survival time will be estimable using the area
    under the Kaplan-Meier curve, and, if desired, plots of power versus design 
    parameters and empirical power based on simulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anneae/RMSTdesign}
  \item Report bugs at \url{https://github.com/anneae/RMSTdesign/issues}
}

}
\author{
\strong{Maintainer}: Anne Eaton \email{eato0055@umn.edu} (0000-0001-5900-9745)

}
\keyword{internal}
