% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSTpow.R
\name{RMSTpow}
\alias{RMSTpow}
\title{Sample Size and Power for the Test of the Difference in Restricted Mean Survival Time}
\usage{
RMSTpow(survdefC, survdefT, k1, k2, tau, n = NA, power = NA,
  plot = F, sim = F, M = 1000, method = "tau_star",
  alpha = 0.025)
}
\arguments{
\item{survdefC}{the survival distribution of the control group, as a list in the form output by \code{survdef}.}

\item{survdefT}{the survival distribution of the treatment group, as a list in the form output by \code{survdef}.}

\item{k1}{length of the accrual period. We assume subjects will accrue
uniformly over the interval \code{(0, k1)} and then be followed until trial time \code{k1+k2}.}

\item{k2}{length of the follow-up period.}

\item{tau}{restriction time for the RMST.}

\item{n}{total sample size for both groups. 1:1 randomization is assumed.
Either \code{n} or \code{power} can be specified, and the other value will be calculated.}

\item{power}{the desired power.}

\item{plot}{if T, plots of power versus each design parameter will be produced. Default is F.}

\item{sim}{if T, simulations will be conducted and empirical power and other
summary statistics will be provided. Default is F. The hypothesis tests are
carried out based on the normal approximation with the variance estimated
according to the Greenwood plug-in/infinitesimal jackknife method.
Specifying situations where survival doesn't go to zero in a reasonable
amount of time (trial length times 1000) will lead to problems if the
\code{sim=T} option is used.}

\item{M}{number of simualations. Default is 1000.}

\item{method}{modification to be used in simulations if the Kaplan-Meier
estimator is not defined at time \code{tau} in either group. Default is 'tau_star',
which changes the restriction time to the last censoring time,
if the last observation is censored at a time earlier than \code{tau}.
Other possible values are 'gill', 'efron', 'tau_star', 'risk1', 'risk2',
and 'risk5'. The riskX' options indicate estimating RMST difference at the
latest time at which at least X people are at risk in each group,
irrespective of the value of \code{tau}.}

\item{alpha}{one-sided type I error level. Default is 0.025.}
}
\value{
a list with components
\item{n}{the user specified n, or if n was left blank, the n needed to achieve the user-specified power.}
\item{powerRMST}{the user specified power, or if power was left blank, the asymptotic power of the RMST test.}
\item{powerLR}{the asymptotic power of the log-rank test.}
\item{pKME}{the probability that you will be able to estimate RMST difference at time tau using the standard Kaplan-Meier estimator.}
\item{plotvals}{a list used for generating plots, returned if \code{plot = T},
with components:}
\itemize{
\item{\code{n_vec}}{ a vector of sample sizes near the selected sample size}
\item{\code{n_pow}}{ power of the RMST based test for the values in n_vec, holding k1, k2 and tau constant  }
\item{\code{k1_vec}}{ a vector of accrual periods near the specified k1}
\item{\code{k1_pow}}{ power of the RMST based test for the values in k1_vec, holding n, tau and total trial length constant}
\item{\code{k2_vec}}{ a vector of follow-up periods near the specified k2}
\item{\code{k2_pow}}{ power of the RMST based test for the values in k2_vec, holding n, k1 and tau constant  }
\item{\code{tau_vec}}{ a vector of restriction times near the specified tau}
\item{\code{tau_pow}}{ power of the RMST based test for the values in tau_vec, holding n, k1 and k2 constant}
}
\item{simout}{a list returned if \code{sim = T}, with components:}

\itemize{
\item{\code{emppowRMST}}{ empirical power of the RMST test.}
\item{\code{emppowLR}}{ empirical power of the log-rank test.}
\item{\code{emppKME}}{ proportion of simulations where the standard KM estimator was used.}
\item{\code{meandiff}}{ mean estimated difference in RMST across the simulated datasets.}
\item{\code{SDdiff}}{ standard deviation of the estimated difference in RMST across the simulated datasets.}
\item{\code{meantrunc}}{ mean truncation time used in the simulated datasets (may be smaller than tau if method = 'tau_star' or 'riskX' options are used).}
\item{\code{SDtrunc}}{ standard deviation of the truncation time used in the simulated datasets.}
}
}
\description{
Determine the asmymptotic power of the test of RMST under a given trial design, or
calculate the samples size needed to achieve a desired power.
}
\examples{
con<-survdef(times = 3, surv = 0.5)
trt<-survdef(haz = 0.67*con$h(1))
RMSTpow(con, trt, k1 = 0, k2 = 3, tau = 3, power = 0.8)
RMSTpow(con, trt, k1 = 0, k2 = 3, tau = 3, n = 552)
}
