% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_duration.R
\name{shortest_duration}
\alias{shortest_duration}
\title{Find Shortest Duration Trial}
\usage{
shortest_duration(survdefC, survdefT, tau, power, accrual_rate,
  pKME = 0.95, alpha = 0.025, altdesign = F, multiplier = 1.1)
}
\arguments{
\item{survdefC}{the survival distribution of the control group, as a list in the form output by survdef.}

\item{survdefT}{the survival distribution of the treatment group, as a list in the form output by survdef.}

\item{tau}{restriction time for the RMST.}

\item{power}{the desired power.}

\item{accrual_rate}{the planned accrual rate, per unit of time.}

\item{pKME}{The desired probability that the RMST difference will be
estimable using the Kaplan-Meier estimator. Default is 0.95.}

\item{alpha}{one-sided type I error level. Default is 0.025.}

\item{altdesign}{if T, the function will look for an alternative design that
is not the shortest in duration, but has duration equal to some multiple of
the shortest possible duration. Default is F. The sample size of the shortest
duration trial can be much larger than a slightly longer trial, so we
recommend considering an alternative design slightly longer than the
shortest trial in addition to the shortest trial.}

\item{multiplier}{the factor by which the duration of the shortest possible
trial is multiplied to acquire the duration of the alternative trial design.
Default is 1.1, meaning a trial that is 10% longer in duration than the
shortest possible trial. This argument is ignored if altdesign=F.}
}
\description{
Find the trial with the shortest duration in calendar time with a specified
power and probability that RMST difference will be estimable with the
Kaplan-Meier estimator. Within all trials of that minimum duration, the
function finds the one with the smallest sample size.
}
\examples{
con<-survdef(times = 3, surv = 0.5)
trt<-survdef(haz = 0.67*con$h(1))
shortest_duration(con, trt, 3, .8, 552/4)
}
