% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdefHR.R
\name{survdefHR}
\alias{survdefHR}
\title{Create a Survdef Object Based on a Hazard Ratio}
\usage{
survdefHR(survdefC, HR)
}
\arguments{
\item{survdefC}{the survival distribution for the reference/control group,
as a list in the form output by \code{survdef}.}

\item{HR}{the hazard ratio defining the relationship between the two distributions.}
}
\value{
a list with components:
\item{S}{a vectorized function that takes time as input and returns the survival probability at that time}
\item{h}{a vectorized function that takes time as input and returns the hazard at that time}
}
\description{
Creates a new object which stores user-specified survival distribution
information in the format needed for the main function, \code{RMSTpow}.
\code{survdefHR} is used when the user wishes to specify a survival distribution
that is defined by its relationship to another distribution via a constant
hazard ratio.
}
\examples{
con<-survdef(times = 3, surv = 0.5); survdefHR(con, 0.5)
}
