% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdefWeibull.R
\name{survdefWeibull}
\alias{survdefWeibull}
\title{Create a Survdef Object for a Weibull Distribution}
\usage{
survdefWeibull(shape, scale)
}
\arguments{
\item{shape}{the shape parameter for the Weibull distribution.
Parametrization is according to built-in R functions for the Weibull
distribution, see \code{?Weibull} for more information.}

\item{scale}{the scale parameter for the Weibull distribution.}
}
\value{
a list with components:
\item{S}{a vectorized function that takes time as input and returns the survival probability at that time}
\item{h}{a vectorized function that takes time as input and returns the hazard at that time}
}
\description{
Create a Survdef Object for a Weibull Distribution
}
\examples{
survdefWeibull(shape = 1.05, scale = 8573)
}
